requirements.txt
parse
pycryptodome

